Fix mis-translation by babel
Babel was turning the "continue" statements in the for-of loop into
"return" statements.  If this were inside an anonymous function, this
would make sense, but it was not.  This would cause certain inputs to
cause a return of undefined in the compiled bundle.

This restructures the source to use if/else if/else instead of
continue in the loop.

Closes #3